move CurtinCommandRunner to its own file
